Create 'fix-list-width' behavior & use it for lists on worklists, schedule, pt flow, pt dashboard/archive
