+++ src/js/views/patients/schedule/schedule_views.js
@@ -8,6 +8,8 @@ import { alphaSort } from 'js/utils/sorting';
 import intl from 'js/i18n';
 import buildMatchersArray from 'js/utils/formatting/build-matchers-array';
 
+import FixListWidthBehavior from 'js/behaviors/fix-list-width';
+
 import 'scss/modules/buttons.scss';
 import 'scss/modules/list-pages.scss';
 import 'scss/modules/table-list.scss';
@@ -26,6 +28,7 @@ import './schedule-list.scss';
 const LayoutView = View.extend({
   className: 'flex-region',
   template: LayoutTemplate,
+  behaviors: [FixListWidthBehavior],
   regions: {
     filters: '[data-filters-region]',
     table: '[data-table-region]',
@@ -43,31 +46,19 @@ const LayoutView = View.extend({
     count: '[data-count-region]',
   },
   childViewEvents: {
-    'update:listDom': 'fixWidth',
+    'update:listDom': 'fixListWidth',
   },
   ui: {
     listHeader: '.js-list-header',
     list: '.js-list',
   },
-  initialize() {
-    const userActivityCh = Radio.channel('user-activity');
-    this.listenTo(userActivityCh, 'window:resize', this.fixWidth);
-  },
-  fixWidth() {
-    /* istanbul ignore if */
-    if (!this.isRendered()) return;
-
-    const headerWidth = this.ui.listHeader.width();
-    const listWidth = this.ui.list.contents().width();
-    const listPadding = parseInt(this.ui.list.css('paddingRight'), 10);
-    const scrollbarWidth = headerWidth - listWidth;
-
-    this.ui.list.css({ paddingRight: `${ listPadding - scrollbarWidth }px` });
-  },
   templateContext() {
     return {
       isReduced: this.getOption('isReduced'),
     };
+  },
+  fixListWidth() {
+    this.triggerMethod('fix:list:width');
   },
 });
