+++ src/js/views/patients/worklist/worklist_views.js
@@ -7,6 +7,8 @@ import intl, { renderTemplate } from 'js/i18n';
 import underscored from 'js/utils/formatting/underscored';
 import buildMatchersArray from 'js/utils/formatting/build-matchers-array';
 
+import FixListWidthBehavior from 'js/behaviors/fix-list-width';
+
 import 'scss/modules/buttons.scss';
 import 'scss/modules/list-pages.scss';
 import 'scss/modules/table-list.scss';
@@ -29,6 +31,7 @@ const i18n = intl.patients.worklist.worklistViews;
 const LayoutView = View.extend({
   className: 'flex-region',
   template: LayoutTemplate,
+  behaviors: [FixListWidthBehavior],
   regions: {
     dateFilter: '[data-date-filter-region]',
     filters: '[data-filters-region]',
@@ -48,7 +51,7 @@ const LayoutView = View.extend({
     count: '[data-count-region]',
   },
   childViewEvents: {
-    'update:listDom': 'fixWidth',
+    'update:listDom': 'fixListWidth',
   },
   triggers: {
     'click @ui.select': 'click:select',
@@ -58,20 +61,8 @@ const LayoutView = View.extend({
     list: '.js-list',
     select: '.js-select',
   },
-  initialize() {
-    const userActivityCh = Radio.channel('user-activity');
-    this.listenTo(userActivityCh, 'window:resize', this.fixWidth);
-  },
-  fixWidth() {
-    /* istanbul ignore if */
-    if (!this.isRendered()) return;
-
-    const headerWidth = this.ui.listHeader.width();
-    const listWidth = this.ui.list.contents().width();
-    const listPadding = parseInt(this.ui.list.css('paddingRight'), 10);
-    const scrollbarWidth = headerWidth - listWidth;
-
-    this.ui.list.css({ paddingRight: `${ listPadding - scrollbarWidth }px` });
+  fixListWidth() {
+    this.triggerMethod('fix:list:width');
   },
 });
